timeout still not working

+++ signals.cpp
@@ -65,10 +65,13 @@ void ctrlCHandler(int sig_num) {
 }
 
 void alarmHandler(int sig_num) {
-  std::cout<<"im in "<<std::endl;
-  // TODO: Add your implementation
  SmallShell &smash = SmallShell::getInstance();
   pid_t procc_pid = smash.getAlarmsList()->getLastAlarmPid();
+  if (procc_pid == -1)
+  {
+    return;
+  }
+  smash.getAlarmsList()->removeLastAlarm();
   if (kill(procc_pid,0) == -1 ) // in case the process has already ended
   {
     int res =waitpid(procc_pid,nullptr,WNOHANG); // kill Zombie;
@@ -79,7 +82,9 @@ void alarmHandler(int sig_num) {
   {
     CALL_SYS(kill(procc_pid,SIGKILL),"kill");
     int res =waitpid(procc_pid,nullptr,WNOHANG); // kill Zombie;
+    std::cout << "before"<<std::endl;
     std::string cmd_line=smash.getCurrFgCommand()->getCmdLine();
+    std::cout << "after"<<std::endl;
     if ( smash.getCurrFgPid() == -1 || smash.getCurrFgPid() != procc_pid) // external command
     {
        cmd_line = smash.getJobsList()->getJobByPid(procc_pid)->getJobCommand();
